Adição de testes unitários
