--- tests/SuperLoja.Api.Tests/Domain/ProdutoComMesmoCodigoSpecificationTests.cs
@@ -1,24 +1,0 @@
-﻿using SuperLoja.Api.Domain.Entidades;
-using SuperLoja.Api.Domain.Specs;
-
-namespace SuperLoja.Api.Tests.Builders.Domain;
-
-public class ProdutoComMesmoCodigoSpecificationTests
-{
-    [Fact]
-    public void EhSatisfeito_DeveRetornarTrue_QuandoProdutoComMesmoCodigoExisteNaLista()
-    {
-        // Arrange
-        var lista = new List<Produto>()
-        {
-            new ProdutoBuilder().ComCodigo("codigo").Build()
-        };
-
-        // Act
-        var item = lista.Where(new ProdutoComMesmoCodigoSpecification("codigo").EhSatisfeito)
-            .FirstOrDefault();
-
-        // Assert
-        Assert.NotNull(item);
-    }
-}
